+++ lib/jahuty/data/render.rb
@@ -1,6 +1,6 @@
 module Jahuty
   module Data
-    class Snippet
+    class Render
       attr_accessor :id, :content
 
       def initialize(id, content)
@@ -12,7 +12,7 @@ module Jahuty
         raise ArgumentError.new "Key :id does not exist" if !data.key?(:id)
         raise ArgumentError.new "Key :content does not exist" if !data.key?(:content)
 
-        Snippet.new(data[:id], data[:content])
+        Render.new(data[:id], data[:content])
       end
 
       def to_s