+++ spec/jahuty/snippet_spec.rb
@@ -1,6 +1,6 @@
 module Jahuty
   RSpec.describe Snippet do
-    describe "::get" do
+    describe "::render" do
       before { Jahuty.key = 'kn2Kj5ijmT2pH6ZKqAQyNexUqKeRM4VG6DDgWN1lIcc' }
 
       it "returns content" do